refactored imageblock and latestprofilelist further
