+++ network-api/networkapi/wagtailpages/customblocks/annotated_image_block.py
@@ -1,17 +1,6 @@
-from wagtail.images.blocks import ImageChooserBlock
 from wagtail.core import blocks
 
-
-class ImageBlock(blocks.StructBlock):
-    image = ImageChooserBlock()
-    altText = blocks.CharBlock(
-        required=True,
-        help_text='Image description (for screen readers).'
-    )
-
-    class Meta:
-        icon = 'image'
-        template = 'wagtailpages/blocks/image_block.html'
+from .image_block import ImageBlock
 
 
 class AnnotatedImageBlock(ImageBlock):
